% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaV3.R
\name{InfCriteriaV3}
\alias{InfCriteriaV3}
\title{Calculates IC values}
\usage{
InfCriteriaV3(loglikelihood, clusters, dimension, nObservations = 100)
}
\arguments{
\item{loglikelihood}{Negative value of class numeric}

\item{clusters}{yes}

\item{dimension}{yes}

\item{nObservations}{yes}
}
\value{
List of BIC and AIC
\itemize{
\item BIC
\item AIC
}
}
\description{
Calculates BIC and AIC, Bayesian and Akaike given log-likelihood
}
\examples{
resultsInfCriteria <- InfCriteriaV3(loglikelihood = -5080,
   clusters = 2,
   dimension = 3,
   nObservations = 1000)
}
